requirements.txt
Django~=2.2.4
django-ckeditor
